update: メソッドを別ファイル化
